+++ packages/aws/src/api-gateway/api.ts
@@ -1,7 +1,7 @@
 import * as aws from "@notation/aws.iac";
 
 export const api = (rgConfig: { name: string }) => {
-  const apiGroup = new aws.AwsResourceGroup("api", rgConfig);
+  const apiGroup = new aws.AwsResourceGroup("API Gateway", rgConfig);
 
   const apiResource = apiGroup.add(
     new aws.apiGateway.Api({