+++ packages/aws/src/lambda.fn/handlers.ts
@@ -1,4 +1,11 @@
-import type { ApiGatewayHandler, EventBridgeHandler } from "src/shared/lambda.handler";
+import type {
+  ApiGatewayHandler,
+  DynamoDbBatchHandler,
+  DynamoDbStreamHandler,
+  SqsBatchHandler,
+  SqsHandler,
+  EventBridgeHandler
+} from "src/shared/lambda.handler";
 
 export const handle = {
   apiRequest:
@@ -8,4 +15,20 @@ export const handle = {
   eventBridgeScheduledEvent: 
     (handler: EventBridgeHandler<"Scheduled Event", any>): EventBridgeHandler<"Scheduled Event", any> =>
     async(...args) => handler(...args)
+  dynamoDbStream:
+    (handler: DynamoDbStreamHandler): DynamoDbStreamHandler =>
+    async (...args) =>
+      handler(...args),
+  dynamoDbBatch:
+    (handler: DynamoDbBatchHandler): DynamoDbBatchHandler =>
+    async (...args) =>
+      handler(...args),
+  sqsEvent:
+    (handler: SqsHandler): SqsHandler =>
+    async (...args) =>
+      handler(...args),
+  sqsBatch:
+    (handler: SqsBatchHandler): SqsBatchHandler =>
+    async (...args) =>
+      handler(...args),
 };