+++ packages/dashboard/src/main.tsx
@@ -1,3 +1,4 @@
+import "./main.css";
 import React from "react";
 import ReactDOM from "react-dom/client";
 import App from "./app.tsx";