+++ packages/core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @notation/core
 
+## 0.5.0
+
+### Minor Changes
+
+- 5debdd1: Show deployed resource state in dashboard
+
+## 0.4.1
+
+### Patch Changes
+
+- Removed dev artifacts from dist
+
 ## 0.4.0
 
 ### Minor Changes