+++ packages/create-notation/CHANGELOG.md
@@ -1,5 +1,17 @@
 # create-notation
 
+## 0.0.3
+
+### Patch Changes
+
+- Add hashbang
+
+## 0.0.2
+
+### Patch Changes
+
+- 10a8133: Migrate project scaffolding to create-notation
+
 ## 0.0.1
 
 ### Patch Changes